requirements.txt
Babel==1.3
BeautifulSoup==3.2.0
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.4.0
PyXML==0.8.4
South==0.8.4
Unidecode==0.04.16
certifi==14.05.14
distribute==0.6.10
django-admin-shortcuts==1.2.5
django-allauth==0.16.1
django-appconf==0.6
django-autocomplete-light==2.0.0a18
django-autoslug-field==0.2.3
django-ckeditor==4.0.2
django-ckeditor-updated==4.4.0
django-compressor==1.3
django-debug-toolbar==0.9.4
django-endless-pagination==2.0
django-extra-views==0.6.5
-e git+https://github.com/toastdriven/django-haystack.git@6b67401aea8203fa172ae49a83c2c5356322d8b5#egg=django_haystack-2.2.0-py2.7-dev
django-localflavor==1.0
django-oscar==0.7.1
django-oscar-paypal==0.9.3
django-suit==0.2.11
django-test-pep8==0.1
django-treebeard==2.0
djangocms-admin-style==0.2.2
lxml==3.4.0
oauthlib==0.6.3
pep8==1.5.7
phonenumbers==6.0.0a
purl==1.0
python-openid==2.2.5
pytz==2014.7
requests==2.4.0
requests-oauthlib==0.4.1
six==1.7.3
sorl-thumbnail==11.12
sqlparse==0.1.13
suds==0.4
wsgiref==0.1.2
